Complete renaming of repo to codebin.it from codebin
